Change the default orientation to portrait.
bug:5446617

Since the orientation is mainly portrait on phones, the default
orientation of an application will decide if the window animation will
be played before it starts. In order to reduce the launch time of Camera
apps, this change is to skip the window animations by changing the default
orientation to portrait.

Change-Id: I6682ab408d7e8d1f0580f3c1600b6c9c3d6a7f6e