Camera OMAP4 : Fix for UI Anmination in Potrait timelapse recording
In Timelapse recording, An animation is drawn on the wheel UI. In
Potrait mode, Wheel UI is not present.
This patch make sure the Animation is drawn only when the wheel UI
is present.
Patch set#2 : Modified the commit message and logic added to specific
failure scenario that a generic change.

Signed-off-by: Sathya Prakash M R <a0393677@ti.com>
Signed-off-by: Jean Johnson <a0271255@ti.com>
Change-Id: I3bab78ede211b3a203e399feb782709ed6330313