OMAP4 Camera : Fix for Potrait mode settings for FRONT Facing Camera
Recorded Video in Potrait Mode for Secondary Camera was recorded in
the correct orientation. This was because the Angle of Rotation is different
for Front Facing Camera. For a particular angle, the Front Facing Camera
rotation is counter clockwise.
This patch makes sure the App sets the correct angle based on the Camera.
Also Mirroring was incorrect due to additional setDisplayOrientation() call.

Signed-off-by: Prasanna Kumar M.R. <x0118059@ti.com>
Signed-off-by: Sathya Prakash M R <a0393677@ti.com>
Change-Id: I5c20f927f65452cb1966e2011d90c2cfc1eaf932