Set the orientation of a panorama image.
The panorama image returned from the panorama library is oriented based
on the natural orientation of the camera.
We set an orientation in its EXIF header, so that the image can be
displayed correctly. The orientation is calculated by compensating the
device orientation and the camera orientation respective to the natural
orientation of the device.

Bug: 5603313
Change-Id: Id6907cb08a7ff77e44736f3c3cd1503e7a68d783