OMAP4 Camera: Patch for disabling rotation feature with MR1 update
Video rotation is broken with the MR1 changes from Google where they
have configured the orientation to Landscape mode for Tablet mode.

As part of the existing rotation logic, we lock and unlock the
orientation during the recording. So when we try to set the orientation
while stopping the recording, it fails. Hence we are only removing the
locking and unlocking of orientation during start/stop of recording

The same video rotation logic has been functional on GB and was also
ported to ICS MR0. New changes from google in MR1 have caused this regression.

We are working on getting the video rotation to be functional on MR1.
We might implement lock/unlock in a different way or if this change is
required we will enable it

Change-Id: I249b54da3735801b2c5ff5f31005d49383d70e70
Signed-off-by: Sathya Prakash M R <a0393677@ti.com>
Signed-off-by: Prasanna Kumar M.R. <x0118059@ti.com>