Camera Application OMAP4: Correct S3D aspect ratio
The camera preview aspect ratio needs adjustment
when selecting non subsampled preview resolutions,
for example 1280x1440.

When the S3DView library is notified that the camera preview
is S3D, then for non-subsampled resolutions the destination
window should look the same as an equivalent 2D resolution.

If the S3DView configuration is set to display as mono content,
then the aspect ratio is unmodified to avoid aspect ratio distortion.

Change-Id: I379c0b301760d37b6e1235fd14377201010d00fb
Signed-off-by: Tsvetan Ivanov <x0161127@ti.com>