Camera OMAP4 : Fix App startup for Blaze
With Rotation fixes over MR-1 release, the OMAPCamera App was
unable to open in Blaze. This was due to different Control Panel alignment
being different for Blaze ( considered Phone kind of device ) and Tablet.

This patch Adjusts the Control Panel Alignment based on the kind of device.
With this patch, the App Controls are at the bottom of the App unlike Standard
Camera App.

Change-Id: I7db9b89017d3b0c9b5731ea340a05d75e76daabd
Signed-off-by: Sathya Prakash M R <a0393677@ti.com>