Camera OMAP4: CamCorder framerate menu settings fix
As min and max VFR settings in CamCorder are separate menu items those need to be properly updated
in case min > max.
This fix checks if min > max VFR border and makes both borders equal to the one that was actually changed
by the user in the menu. Menu is updated accordingly.

Change-Id: I63d274d6da152d5ff27e43ed0483da67ad1f929a
Signed-off-by: Pavel Nedev <pnedev@mm-sol.com>