OMAPCamera: Fix for initial orientation is not received.
There is an issue if the initial orientation, that is, mOrientation
value just after lunching the OMAPCcamera app.
If onOrientationChanged() is not called before capture, the
mOrientation will not be updated and the Util.setRotationParameter()
will set the rotation as 0. Due to this the captured images will be
encoded in wrong orientation ( This happens when the sensor mount
orientation is not zero).

This patch will set the rotation parameter with the sensor mount
orientation if the mOrientation = ORIENTATION_UNKNOWN. So that the
captured image will be at correct orientation.

Change-Id: Ic148575a96960462dd64da5bbf3dcb12e6c51c2d
Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>