OMAPCamera:Temp.bracketing: very first orientation is not received.
For Temporal bracketing setparameters() will not be called in capture().
So rotation parameter will not be set before capture.
Due to this, the sensor mount orientation will not be considered unless
onOrientationChanged() get triggered before capture.

So incase, capture is triggered before onOrientationChanged(),
rotation parameter should be set in oder to get the proper
orientation of captured image.
This is valid only if the sensor mount orientation is not zero.

Change-Id: Icba2f021e1a0ad21f70be129879305d22e0459da
Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>