OMAPCamera : Fix Preveiw Orientation in Video mode for New Camera layout
OMAP5 SEVM device is a portrait Device. This affects the Camcorder rotation usecase.
An offset is added to correct the Orientation of preview and recorded clips.
The offset is added only in case of Portrait devices for co-existance with
Landscape devices like Tablets and our Blaze Platform.

With this change, whenever orientation is changed during video preview
startPreview() will be called from onOrientationChanged() to set the
right orientation.

Patch set#4,#5: Fixed a miscalculation resulting in wrong preview angle
for Blaze Tablet and Blaze.

Change-Id: I580c65bf9cac4765b741564468284f814b419e92
Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>
Signed-off-by: Sathya Prakash M R <a0393677@ti.com>